Move brand store components
