+++ static/js/brand-store/brand-store.js
@@ -2,7 +2,7 @@ import React from "react";
 import { createRoot } from "react-dom/client";
 import * as Sentry from "@sentry/react";
 import { Integrations } from "@sentry/tracing";
-import App from "./App";
+import App from "./components/App";
 import { store } from "./store";
 import { Provider } from "react-redux";
