+++ static/js/brand-store/components/Model/index.ts
