+++ static/js/brand-store/components/Publisher/index.ts
