+++ static/js/brand-store/components/Snaps/SnapsTable.js
