+++ static/js/brand-store/components/Snaps/SnapsTableRow.js
